layout for School Selection screen
the basic layout for what the school selections screen will be
it is based off of the already made Login Screen. I commented out functions from the Login Screen that may be used here later once the mock api is integrated with it
This screen uses a new SchoolPicker component, which is a modal component that overlays itself atop the screen and displays a data set of schools. You can tap on a school, and it will be sent back to the main screen for processing.