Added a new Theme to the Card Component
Added both a new theme to the card component, "list", aswell as the "theme?" property.
This new theme displays the card with smaller text and a more compact form, aswell as only displaying the preview if "urgent? = true".
This theme of card is meant to mimic the options in a hamburger menu.

I also added a list of cards to the Records page in Student General Info