created backbone for student schedule screen
Made a new component, Parent_ViewClassModal, which holds information about a student's class relative to the parent who is accessing it. Eventually, this modal will be provided the studentId and classId by a global state variable, but right now it holds placeholder values.
studentSchedule screen displays like the other student-centric general-info parent screens. It holds a list of classes, which will eventually be provided by the same global state variable, but for now is static.